+++ Sources/FluentBenchmark/SolarSystem/Moon.swift
@@ -24,14 +24,12 @@ public final class Moon: Model {
         id: IDValue? = nil,
         name: String,
         craters: Int,
-        comets: Int,
-        planetID: Planet.IDValue
+        comets: Int
     ) {
         self.id = id
         self.name = name
         self.craters = craters
         self.comets = comets
-        self.$planet.id = planetID
     }
 }
 
@@ -57,21 +55,42 @@ public final class MoonSeed: Migration {
     public init() { }
 
     public func prepare(on database: Database) -> EventLoopFuture<Void> {
-        let saves = [
-            Moon(name: "Deimos", craters: 1, comets: 5, planetID: 1),
-            Moon(name: "Prometheus", craters: 8, comets: 19, planetID: 2),
-            Moon(name: "Hydra", craters: 2, comets: 2, planetID: 3),
-            Moon(name: "Luna", craters: 10, comets: 10, planetID: 4),
-            Moon(name: "Atlas", craters: 9, comets: 8, planetID: 5),
-            Moon(name: "Janus", craters: 15, comets: 9, planetID: 6),
-            Moon(name: "Phobos", craters: 20, comets: 3, planetID: 7)
-        ].map { moon -> EventLoopFuture<Void> in
-            return moon.save(on: database)
+        Planet.query(on: database).all().flatMap { planets in
+            .andAllSucceed(planets.map { planet in
+                let moons: [Moon]
+                switch planet.name {
+                case "Earth":
+                    moons = [
+                        .init(name: "Luna", craters: 10, comets: 10)
+                    ]
+                case "Mars":
+                    moons = [
+                        .init(name: "Deimos", craters: 1, comets: 5),
+                        .init(name: "Phobos", craters: 20, comets: 3)
+                    ]
+                case "Jupiter":
+                    moons = [
+                        .init(name: "Io", craters: 10, comets: 10),
+                        .init(name: "Europa", craters: 10, comets: 10),
+                        .init(name: "Ganymede", craters: 10, comets: 10),
+                        .init(name: "callisto", craters: 10, comets: 10),
+                    ]
+                case "Saturn":
+                    moons = [
+                        .init(name: "Titan", craters: 10, comets: 10),
+                        .init(name: "Prometheus", craters: 10, comets: 10),
+                        .init(name: "Atlas", craters: 9, comets: 8),
+                        .init(name: "Janus", craters: 15, comets: 9)
+                    ]
+                default:
+                    moons = []
+                }
+                return planet.$moons.create(moons, on: database)
+            }, on: database.eventLoop)
         }
-        return .andAllSucceed(saves, on: database.eventLoop)
     }
 
     public func revert(on database: Database) -> EventLoopFuture<Void> {
-        return database.eventLoop.makeSucceededFuture(())
+        Moon.query(on: database).delete()
     }
 }
