+++ Sources/FluentBenchmark/Tests/Migrator.swift
@@ -1,0 +1,51 @@
+extension FluentBenchmarker {
+    public func testMigrator() throws {
+        try self.runTest(#function, []) {
+            let migrations = Migrations()
+            migrations.add(GalaxyMigration())
+            migrations.add(PlanetMigration())
+
+            let migrator = Migrator(
+                databaseFactory: { _ in self.database },
+                migrations: migrations,
+                on: self.database.eventLoop
+            )
+            try migrator.setupIfNeeded().wait()
+            try migrator.prepareBatch().wait()
+
+            migrator.migrations.add(GalaxySeed())
+            try migrator.prepareBatch().wait()
+
+            let logs = try MigrationLog.query(on: self.database)
+                .all().wait()
+                .map { $0.batch }
+            XCTAssertEqual(logs, [1, 1, 2], "batch did not increment")
+
+            try migrator.revertAllBatches().wait()
+
+        }
+    }
+
+    public func testMigratorError() throws {
+        try self.runTest(#function, []) {
+            let migrations = Migrations()
+            migrations.add(GalaxyMigration())
+            migrations.add(ErrorMigration())
+            migrations.add(PlanetMigration())
+
+            let migrator = Migrator(
+                databaseFactory: { _ in self.database },
+                migrations: migrations,
+                on: self.database.eventLoop
+            )
+            try migrator.setupIfNeeded().wait()
+            do {
+                try migrator.prepareBatch().wait()
+                XCTFail("Migration should have failed.")
+            } catch {
+                // success
+            }
+            try migrator.revertAllBatches().wait()
+        }
+    }
+}
