+++ Sources/FluentBenchmark/Utilities/SchoolSeed.swift
